Tests/LibWeb/Text/expected/css/fit-content-composition.txt

fit-content(10px)
fit-content(15px)
fit-content(20px)
